Merged PR 34119: Register cookies in the consent manager
## What's being changed

This PR adds cookies used by our app to the Shopware consent manager. The cookies are listed in a group, with snippets to add labels and descriptions in English and German.

We have two cookies that are dropped by remote scripts:
Name | Script | Cookie
| ----------- | ----------- | ----------- |
WBT | //static.trackedweb.net/js/_dmptv4.js | dmSessionID
Site/ROI tracking | //r1-t.trackedlink.net/_dmpt.js | dm_i

We also set an additional cookie - the 'Governor' cookie - which is the one that decides whether or not to load our tracking scripts.

Only if the Governor is accepted will we load our scripts for WBT and Site/ROI.

## Why it's being changed

App submission requires this declaration of cookies for consent by the user.

The Governor cookie is a necessary addition, because our `dmSessionID` and `dm_i` cookies are loaded dynamically via the remote scripts. We can't set a specific value for these cookies in manifest.xml, therefore Shopware doesn't store the consent (it just passes it to the event). So - if you accept the Governor, everything works. If you don't, nothing works. If you opt out of either of the two tracking cookies, we opt you out of everything.

The way in which we check for cookies probably needs some explaining.
- The Site/ROI script is loaded in a twig template, so we need to insert some JavaScript here to prevent it loading at all if the Governor is not set.
- We could do this the same way for the WBT plugin, but it's easier to use `CookieStorage.getItem()` in the plugin itself.
- Once these two checks are in place and working, in the other plugins (roi, identify and cartinsight) we can simply check for `window.dmPt` or `window._dmTrack`.

## How to review / test this change

- Uninstall first
```
bin/console app:uninstall DotdigitalForShopware
```
- Now re-install to load the updated manifest.xml
```
bin/console app:install --activate DotdigitalForShopware
```
- Check that your app configuration is still correct

_Base state_
- Delete existing cookies, reload and click 'Configure'
- No Dotdigital cookies should be checked
- No tracking scripts should fire
- No network requests should be seen to `PageTrack`, `pagevisit` or `cartinsight`

_Accept the Governor_
- Delete the 'cookie-preference' cookie to review your consent
- Accept the 'Dotdigital enabled' cookie
- Everything should now work
- You should re-run end-to-end tests for site tracking, ROI and WBT

_Deny the Governor_
- Delete the 'cookie-preference' cookie to review your consent
- Uncheck the 'Dotdigital enabled' cookie
- All cookies (`dmSessionID`, `dm_i` and `dd-tracking-enabled`) should be deleted
- No more DD network requests should happen
- Reload a page with the `dm_i` query string
- The `dm_i` cookie should not be dropped

_Opt out of dm_i_
- Delete the 'cookie-preference' cookie to review your consent
- Deny the 'Dotdigital tracking' cookie
- All cookies (`dmSessionID`, `dm_i` and `dd-tracking-enabled`) should be deleted
- No m...