Merged PR 34459: Fix array map syntax
## What's being changed

The array syntax used to loop through our cookies and remove them.

## Why it's being changed

Shopware flagged this in the automatic code review. The problem was in this code:
```
Object.keys(DotdigitalCookies).map(key => {
    CookieStorage.removeItem(DotdigitalCookies[key]);
})
```
The error was 'Add a "return" statement to this callback.'

The `{ }` signify a function, which requires a return statement. I could either a) keep `.map()` and remove the curly braces, or (in preference apparently) use `forEach`, either with or without (again, preferred) the braces.

## How to review / test this change

- Accept cookies
- Reset cookie preference
- Uncheck one Dotdigital cookie
- All Dotdigital cookies should be removed and not come back on the next page load

Related work items: #162796