Merged PR 45302: Merged PR 45268: Revert to using window instead of sessionStorage
## What's being changed

We're aligning our cart insight plugin with the simpler version we were forced to use for Shopware 6.4.x.

## Why it's being changed

It's preferable to have as few functional differences as possible between the Shopware 6.4.x code and the 6.5.x code. Plus, the benefit of setting a property of `window` here is that we send one final payload after a cart is emptied, then no further payloads - whereas with `sessionStorage` we carry on sending empty payloads for the duration of the session (i.e. until the tab or window is closed).

## How to review / test this change

- Smoke test cart insight payload for visitor arriving (incognito)
- Check cart insight payload when cart contains items
- Check cart insight sent once only after cart is emptied

## Notes

Merged PR 45268: Revert to using window instead of sessionStorage

`sessionStorage` has been removed from the cart insight plugin and replaced with setting a variable on `window`.

We went with the more elegant `sessionStorage` in our 2.0.1 release, but in Shopware 6.4.x (node v16), we saw an error on compile:
```
ERROR in /Users/alastairmucklow/Sites/shopware/development/custom/apps/DotdigitalForShopware/Resources/app/storefront/src/plugins/cart-insight-handler.plugin.js 158:63
	Module parse failed: Unexpected token (158:63)
	File was processed with these loaders:
	 * ../node_modules/babel-loader/lib/index.js
	You may need an additional loader to handle the result of these loaders.
	|     key: "cartPreviouslyHadItems",
	|     get: function get() {
	>       return sessionStorage.getItem('cartPreviouslyHadItems') ?? null;
	|     },
	|     set: function set(value) {
```

- In a local 6.4.x instance, check out this branch
- Run `./psh.phar storefront:build`
- See no errors
- Re-run test steps for 'Prevent empty cart payloads for new carts' PR

Related work items: #211863

Related work items: #211863